refactor: extract report events from AgentWorkspace
